requirements.txt
arrow==1.2.1
astroid==2.9.3
autopep8==1.6.0
cachelib==0.6.0
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.0.4
colorama==0.4.4
coverage==6.3
Flask==2.0.3
Flask-Session==0.4.0
idna==3.3
isort==5.10.1
itsdangerous==2.1.1
Jinja2==3.0.3
jinja2-time==0.2.0
lazy-object-proxy==1.7.1
make==0.1.6.post2
MarkupSafe==2.0.1
mccabe==0.6.1
platformdirs==2.4.1
pycodestyle==2.8.0
pylint==2.12.2
python-dateutil==2.8.2
python-dotenv==0.19.2
requests==2.27.1
six==1.16.0
spotipy==2.19.0
toml==0.10.2
urllib3==1.26.8
Werkzeug==2.0.3
wrapt==1.13.3
arrow==1.2.1
astroid==2.9.3
autopep8==1.6.0
cachelib==0.6.0
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.0.4
colorama==0.4.4
coverage==6.3
Flask==2.0.3
Flask-Session==0.4.0
gunicorn==20.1.0
idna==3.3
isort==5.10.1
itsdangerous==2.1.1
Jinja2==3.0.3
jinja2-time==0.2.0
lazy-object-proxy==1.7.1
make==0.1.6.post2
MarkupSafe==2.0.1
mccabe==0.6.1
platformdirs==2.4.1
pycodestyle==2.8.0
pylint==2.12.2
python-dateutil==2.8.2
python-dotenv==0.19.2
requests==2.27.1
six==1.16.0
spotipy==2.19.0
toml==0.10.2
urllib3==1.26.8
Werkzeug==2.0.3
wrapt==1.13.3
